fix(watcher): harden snapshot reconciliation
