+++ test/serve/watch-snapshot.test.ts
@@ -359,6 +359,352 @@ describe("buildWatcherSnapshot + diffWatcherSnapshot", () => {
     expect(built.snapshot.directories.get("")?.has("keep.md")).toBe(true);
   });
 
+  test("missing collection root is scan_failed, not mass deletion", async () => {
+    await write(root, "keep.md", "k");
+    await write(root, "nested/a.md", "a");
+    const built = await buildWatcherSnapshot(root);
+    expect(built.status).toBe("ok");
+    if (built.status !== "ok") {
+      return;
+    }
+
+    const missingRootFs: WatcherSnapshotFs = {
+      readdir: async () => {
+        const error = new Error("ENOENT");
+        (error as NodeJS.ErrnoException).code = "ENOENT";
+        throw error;
+      },
+      lstat: async () => {
+        const error = new Error("ENOENT");
+        (error as NodeJS.ErrnoException).code = "ENOENT";
+        throw error;
+      },
+    };
+
+    const resolved = await resolveWatcherDirtyDirectory(root, "keep.md", {
+      fs: missingRootFs,
+    });
+    expect(resolved.status).toBe("fallback");
+    if (resolved.status === "fallback") {
+      expect(resolved.reason).toBe("scan_failed");
+    }
+
+    const diff = await diffWatcherSnapshot(root, built.snapshot, [""], {
+      fs: missingRootFs,
+    });
+    expect(diff.status).toBe("fallback");
+    if (diff.status !== "fallback") {
+      return;
+    }
+    expect(diff.reason).toBe("scan_failed");
+    // No nextSnapshot field on fallback — proven state is caller-owned and intact.
+    expect(built.snapshot.directories.get("")?.has("keep.md")).toBe(true);
+    expect(built.snapshot.directories.get("nested")?.has("a.md")).toBe(true);
+
+    const reconciled = await reconcileWatcherHints(
+      root,
+      built.snapshot,
+      ["keep.md", "nested/a.md"],
+      { fs: missingRootFs }
+    );
+    expect(reconciled.status).toBe("fallback");
+    if (reconciled.status !== "fallback") {
+      return;
+    }
+    expect(reconciled.reason).toBe("scan_failed");
+  });
+
+  test("millisecond-only timestamps are unreliable_metadata", async () => {
+    await write(root, "a.md", "a");
+    const built = await buildWatcherSnapshot(root);
+    expect(built.status).toBe("ok");
+    if (built.status !== "ok") {
+      return;
+    }
+
+    const msOnlyFs: WatcherSnapshotFs = {
+      readdir: async (absPath: string) => {
+        const { readdir } = await import("node:fs/promises");
+        return readdir(absPath);
+      },
+      lstat: async (absPath: string): Promise<WatcherSnapshotStat> => {
+        const { lstat } = await import("node:fs/promises");
+        const stat = await lstat(absPath, { bigint: true });
+        return {
+          isFile: () => stat.isFile(),
+          isDirectory: () => stat.isDirectory(),
+          isSymbolicLink: () => stat.isSymbolicLink(),
+          dev: stat.dev,
+          ino: stat.ino,
+          size: stat.size,
+          // Intentionally omit mtimeNs/ctimeNs — only ms present.
+          mtimeMs: Number(stat.mtimeNs) / 1_000_000,
+          ctimeMs: Number(stat.ctimeNs) / 1_000_000,
+        };
+      },
+    };
+
+    const diff = await diffWatcherSnapshot(root, built.snapshot, [""], {
+      fs: msOnlyFs,
+    });
+    expect(diff.status).toBe("fallback");
+    if (diff.status !== "fallback") {
+      return;
+    }
+    expect(diff.reason).toBe("unreliable_metadata");
+  });
+
+  test("directory→outside-symlink swap mid-scan fails closed without outside lstat", async () => {
+    await mkdir(join(root, "victim"), { recursive: true });
+    await write(root, "victim/inside.md", "in");
+    await write(root, "root.md", "r");
+
+    const built = await buildWatcherSnapshot(root);
+    expect(built.status).toBe("ok");
+    if (built.status !== "ok") {
+      return;
+    }
+
+    const outsideDir = await mkdtemp(join(tmpdir(), "gno-watch-swap-out-"));
+    const outsideSecret = join(outsideDir, "secret.md");
+    await writeFile(outsideSecret, "secret");
+
+    const { lstat, readdir, rm } = await import("node:fs/promises");
+    const realLstat = async (absPath: string): Promise<WatcherSnapshotStat> => {
+      const stat = await lstat(absPath, { bigint: true });
+      return {
+        isFile: () => stat.isFile(),
+        isDirectory: () => stat.isDirectory(),
+        isSymbolicLink: () => stat.isSymbolicLink(),
+        dev: stat.dev,
+        ino: stat.ino,
+        size: stat.size,
+        mtimeNs: stat.mtimeNs,
+        ctimeNs: stat.ctimeNs,
+      };
+    };
+
+    const victimAbs = join(root, "victim");
+    let victimPreLstats = 0;
+    const childLstats: string[] = [];
+
+    const racingFs: WatcherSnapshotFs = {
+      readdir: async (absPath: string) => {
+        if (absPath === victimAbs) {
+          // After parent listed `victim` as a directory, replace it with an
+          // outside symlink before this directory's own enumeration completes.
+          await rm(victimAbs, { recursive: true, force: true });
+          await symlink(outsideDir, victimAbs, "dir");
+        }
+        return readdir(absPath);
+      },
+      lstat: async (absPath: string): Promise<WatcherSnapshotStat> => {
+        if (absPath === victimAbs) {
+          victimPreLstats += 1;
+          // First lstat (pre-readdir identity): still a real directory.
+          // Subsequent lstat (post-readdir): sees the outside symlink.
+          if (victimPreLstats === 1) {
+            return {
+              isFile: () => false,
+              isDirectory: () => true,
+              isSymbolicLink: () => false,
+              dev: 1n,
+              ino: 42n,
+              size: 64,
+              mtimeNs: 1_000n,
+              ctimeNs: 1_000n,
+            };
+          }
+          return realLstat(absPath);
+        }
+        // Any child under the swapped path would be outside the collection.
+        if (
+          absPath === outsideSecret ||
+          absPath.startsWith(outsideDir + "/") ||
+          absPath.startsWith(outsideDir + "\\")
+        ) {
+          childLstats.push(absPath);
+          return realLstat(absPath);
+        }
+        // Also catch joins that resolve through the symlink path string.
+        if (absPath.includes("secret.md")) {
+          childLstats.push(absPath);
+        }
+        return realLstat(absPath);
+      },
+    };
+
+    try {
+      // Dirty the nested directory so enumeration runs on the race target
+      // (unchanged root fingerprints would not recurse into `victim`).
+      const diff = await diffWatcherSnapshot(root, built.snapshot, ["victim"], {
+        fs: racingFs,
+      });
+      expect(diff.status).toBe("fallback");
+      if (diff.status !== "fallback") {
+        return;
+      }
+      expect(diff.reason).toBe("scan_failed");
+      // Fail closed: never lstat outside children after the identity change.
+      expect(childLstats).toEqual([]);
+      // Proven snapshot is not advanced (no nextSnapshot on fallback).
+      expect(built.snapshot.directories.get("victim")?.has("inside.md")).toBe(
+        true
+      );
+    } finally {
+      await safeRm(outsideDir);
+    }
+  });
+
+  test("many-directory scans stay linear in readdir and lstat operations", async () => {
+    // Pure in-memory tree: N sibling directories each with one file.
+    // Linear bound: readdir once per visited dir; lstat = 2*dirs (identity)
+    // + children. Quadratic map recount / queue.shift would not inflate I/O,
+    // but we still assert tight linear op ceilings so a regression to
+    // re-walking every map between dirs (if reintroduced as I/O) fails.
+    const dirCount = 2_500;
+    type Node =
+      | { kind: "dir"; children: Map<string, Node>; dev: bigint; ino: bigint }
+      | {
+          kind: "file";
+          size: number;
+          dev: bigint;
+          ino: bigint;
+          mtimeNs: bigint;
+          ctimeNs: bigint;
+        };
+
+    let nextIno = 1n;
+    const rootNode: Extract<Node, { kind: "dir" }> = {
+      kind: "dir",
+      children: new Map(),
+      dev: 1n,
+      ino: nextIno,
+    };
+    nextIno += 1n;
+
+    for (let i = 0; i < dirCount; i += 1) {
+      const name = `d${String(i).padStart(4, "0")}`;
+      const dir: Extract<Node, { kind: "dir" }> = {
+        kind: "dir",
+        children: new Map(),
+        dev: 1n,
+        ino: nextIno,
+      };
+      nextIno += 1n;
+      dir.children.set("f.md", {
+        kind: "file",
+        size: 1,
+        dev: 1n,
+        ino: nextIno,
+        mtimeNs: 1_000n,
+        ctimeNs: 1_000n,
+      });
+      nextIno += 1n;
+      rootNode.children.set(name, dir);
+    }
+
+    const resolveNode = (absPath: string): Node | null => {
+      if (absPath === "/mem") {
+        return rootNode;
+      }
+      if (!absPath.startsWith("/mem/")) {
+        return null;
+      }
+      const parts = absPath.slice("/mem/".length).split("/").filter(Boolean);
+      let cursor: Node = rootNode;
+      for (const part of parts) {
+        if (cursor.kind !== "dir") {
+          return null;
+        }
+        const next = cursor.children.get(part);
+        if (!next) {
+          return null;
+        }
+        cursor = next;
+      }
+      return cursor;
+    };
+
+    let readdirCalls = 0;
+    let lstatCalls = 0;
+
+    const memFs: WatcherSnapshotFs = {
+      readdir: async (absPath: string) => {
+        readdirCalls += 1;
+        const node = resolveNode(absPath);
+        if (!node || node.kind !== "dir") {
+          const error = new Error("ENOENT");
+          (error as NodeJS.ErrnoException).code = "ENOENT";
+          throw error;
+        }
+        return [...node.children.keys()];
+      },
+      lstat: async (absPath: string): Promise<WatcherSnapshotStat> => {
+        lstatCalls += 1;
+        const node = resolveNode(absPath);
+        if (!node) {
+          const error = new Error("ENOENT");
+          (error as NodeJS.ErrnoException).code = "ENOENT";
+          throw error;
+        }
+        if (node.kind === "dir") {
+          return {
+            isFile: () => false,
+            isDirectory: () => true,
+            isSymbolicLink: () => false,
+            dev: node.dev,
+            ino: node.ino,
+            size: 64,
+            mtimeNs: 1_000n,
+            ctimeNs: 1_000n,
+          };
+        }
+        return {
+          isFile: () => true,
+          isDirectory: () => false,
+          isSymbolicLink: () => false,
+          dev: node.dev,
+          ino: node.ino,
+          size: node.size,
+          mtimeNs: node.mtimeNs,
+          ctimeNs: node.ctimeNs,
+        };
+      },
+    };
+
+    const built = await buildWatcherSnapshot("/mem", { fs: memFs });
+    expect(built.status).toBe("ok");
+    if (built.status !== "ok") {
+      return;
+    }
+    // Root has dirCount dirs; each subdir has 1 file → 2 * dirCount entries.
+    expect(built.snapshot.entryCount).toBe(dirCount * 2);
+
+    // Visited directories: root + dirCount children.
+    const visitedDirs = dirCount + 1;
+    expect(readdirCalls).toBe(visitedDirs);
+    // Per visited dir: pre-lstat + post-lstat; plus one lstat per child name.
+    // Root children: dirCount; each subdir children: 1 → dirCount files.
+    const expectedLstats = 2 * visitedDirs + dirCount + dirCount;
+    expect(lstatCalls).toBe(expectedLstats);
+
+    // Unchanged root diff must also stay linear (not re-scan every subtree).
+    readdirCalls = 0;
+    lstatCalls = 0;
+    const diff = await diffWatcherSnapshot("/mem", built.snapshot, [""], {
+      fs: memFs,
+    });
+    expect(diff.status).toBe("ok");
+    if (diff.status !== "ok") {
+      return;
+    }
+    expect(diff.candidates).toEqual([]);
+    // Unchanged fingerprints: only the dirty root is enumerated.
+    expect(readdirCalls).toBe(1);
+    expect(lstatCalls).toBe(2 + dirCount);
+  });
+
   test("one changed file among 5000 selects only that path with timing", async () => {
     const total = 5000;
     const writes: Promise<void>[] = [];
